# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: VLTONE Platform API
  version: 1.0.0
  description: Source-of-truth contract shared by the public site, admin site, and desktop application.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/desktop/auth/login:
    post:
      operationId: desktopLogin
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/DesktopLoginRequest" } } }
      responses: { "200": { $ref: "#/components/responses/DesktopSession" }, "409": { $ref: "#/components/responses/Error" } }
  /v1/desktop/auth/refresh:
    post:
      operationId: desktopRefresh
      parameters: [{ name: Idempotency-Key, in: header, required: false, schema: { type: string, format: uuid }, description: Durable refresh operation ID; resend the same ID after a lost response }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/DesktopRefreshRequest" } } }
      responses: { "200": { $ref: "#/components/responses/DesktopSession" }, "401": { $ref: "#/components/responses/Error" } }
  /v1/desktop/auth/logout:
    post:
      operationId: desktopLogout
      security: [{ desktopBearer: [] }]
      responses: { "204": { description: Revoked } }
  /v1/desktop/me:
    get:
      operationId: desktopMe
      security: [{ desktopBearer: [] }]
      responses: { "200": { $ref: "#/components/responses/AccountSession" } }
...
components:
  securitySchemes:
    webCookie: { type: apiKey, in: cookie, name: vlt_web_session }
    adminCookie: { type: apiKey, in: cookie, name: vlt_admin_session }
    csrf: { type: apiKey, in: header, name: X-CSRF-Token }
    desktopBearer: { type: http, scheme: bearer, bearerFormat: Ed25519 }
    reporterBearer: { type: http, scheme: bearer, bearerFormat: opaque }
...
  responses:
    Error:
      description: Structured API error
      content: { application/json: { schema: { $ref: "#/components/schemas/APIError" } } }
    AccountSession:
      description: Authenticated account state
      content: { application/json: { schema: { $ref: "#/components/schemas/AccountSession" } } }
    DesktopSession:
      description: Rotating desktop credentials and offline entitlement
      content: { application/json: { schema: { $ref: "#/components/schemas/DesktopSession" } } }
  schemas:
...
    APIError:
      type: object
      required: [code, message, request_id]
      properties:
        code: { type: string }
        message: { type: string }
        field_errors: { type: object, additionalProperties: { type: string } }
        request_id: { type: string }
...
    User:
      type: object
      required: [id, email, nickname, locale, status, collaboration_enabled]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        nickname: { type: string }
        locale: { type: string, enum: [ru, en] }
        status: { type: string, enum: [active, suspended] }
        collaboration_enabled: { type: boolean }
        consent_version: { type: string }
        consent_accepted_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    Quota:
      type: object
      required: [base_limit, adjustment, used_tokens, reserved_tokens, remaining_tokens, starts_at, ends_at]
      properties:
        base_limit: { type: integer, format: int64 }
        adjustment: { type: integer, format: int64 }
        used_tokens: { type: integer, format: int64 }
        reserved_tokens: { type: integer, format: int64 }
        remaining_tokens: { type: integer, format: int64 }
        starts_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time }
...
    HardwareSnapshot:
      type: object
      additionalProperties: false
      properties:
        cpu_model: { type: string }
        cpu_cores: { type: integer }
        cpu_threads: { type: integer }
        gpu:
          type: array
          maxItems: 8
          items:
            type: object
            additionalProperties: false
            properties: { name: { type: string }, driver: { type: string } }
        ram_bytes: { type: integer, format: int64 }
        os_name: { type: string }
        os_version: { type: string }
        arch: { type: string }
    AccountSession:
      type: object
      required: [user, subscription, quota]
      properties:
        user: { $ref: "#/components/schemas/User" }
        csrf_token: { type: string }
        subscription:
          type: object
          properties: { plan: { const: demo }, display_name: { const: Demo }, all_features: { const: true } }
        quota: { $ref: "#/components/schemas/Quota" }
    DesktopLoginRequest:
      type: object
      additionalProperties: false
      required: [email, password, installation_id, device_name, platform, os_version, app_version, hardware]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 12, maxLength: 128 }
        installation_id: { type: string, format: uuid, description: Random installation UUID; never hardware-derived }
        device_name: { type: string }
        platform: { type: string, enum: [windows, macos] }
        os_version: { type: string }
        app_version: { type: string }
        hardware: { $ref: "#/components/schemas/HardwareSnapshot" }
    DesktopRefreshRequest:
      type: object
      additionalProperties: false
      required: [refresh_token, app_version]
      properties:
        refresh_token: { type: string }
        request_id: { type: string, format: uuid, description: Persist this random ID before refreshing and reuse it after a lost reply until the new credential is durably stored }
        app_version: { type: string }
    DesktopSession:
      allOf:
        - { $ref: "#/components/schemas/AccountSession" }
        - type: object
          required: [access_token, refresh_token, reporter_token, reporter_expires_at, offline_entitlement, server_time, public_key]
          properties:
            access_token: { type: string }
            refresh_token: { type: string }
            reporter_token: { type: string }
            reporter_expires_at: { type: string, format: date-time }
            offline_entitlement: { type: string }
            server_time: { type: string, format: date-time }
            public_key: { type: string }
...